backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
redis==5.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0 
dspy-ai== 2.6.27
PyJWT==2.10.1
tabulate==0.9.0
faker==37.5.3
